refactor: Cleaner scheduler architecture with seperation of concerns between files (also many minor bug fixes)
