+++ src/web/routes/api/system.py
@@ -65,6 +65,7 @@ class SchedulerSummaryModel(BaseModel):
     most_recent_sync: str | None = None
     most_recent_sync_profile: str | None = None
     next_database_sync_at: str | None = None
+    coordinator: dict | None = None
     profiles: dict[str, ProfileStatusModel]
 
 
@@ -126,12 +127,14 @@ async def api_about() -> AboutResponse:
     """
     scheduler = get_app_state().scheduler
     status: dict[str, Any] = {}
+    scheduler_runtime_metrics: dict[str, Any] = {}
     scheduler_running = False
     next_db_sync_iso: str | None = None
 
     if scheduler:
         try:
             status = await scheduler.get_status()
+            scheduler_runtime_metrics = await scheduler.get_runtime_metrics()
             scheduler_running = scheduler.is_running
             next_db_sync = scheduler.get_next_database_sync_at()
             if next_db_sync is not None:
@@ -232,6 +235,7 @@ async def api_about() -> AboutResponse:
         most_recent_sync=most_recent_sync_iso,
         most_recent_sync_profile=most_recent_sync_profile,
         next_database_sync_at=next_db_sync_iso,
+        coordinator=scheduler_runtime_metrics.get("coordinator"),
         profiles=converted,
     )
