+++ src/web/routes/webhook/provider.py
@@ -48,8 +48,8 @@ async def provider_webhook(
                 library_keys,
             )
             schedule_task(
-                scheduler.trigger_sync(
-                    profile_name=profile_name,
+                scheduler.trigger_profile_sync(
+                    profile_name,
                     poll=False,
                     library_keys=library_keys,
                     source="webhook:provider",
